Add vim-rhubarh, reenabled netrw
I'm completely in the dark as to why netrw had been disabled this whole
time in my vimrc; it's _probably_ just that I copied it from someone
else's without understanding why.

In any case, reenabling it allows fugitive's `:GBlame` to actually work
now (this was the motivation for the change in the first place).

Interestingly, in reenableing netrw, doing a fuzzy-find for files with
fzf now also displays a preview of each file, which I can only assume
was also previously broken due to the disabled netrw.

Who knows what else will spring to life now? Very exciting times.