Comment this out for now
The error is saying that python is an unknown argument and therefore
invalid. This wasn't always the case but I can't be bothered right now
to track down what changed where.

I took a very brief look at https://github.com/LnL7/nix-darwin but
nothing jumped out, so my inclination is to comment it out, shrug, and
hope for the best.