Add `dloverflow.sh`
It takes everything older than 30d in `~/Downloads` and moves it
(timestamped) into `~/stale_downloads`, which gets backed up by the
rsync script that runs regularly.

I was really hoping to have this on a schedule too, but there's a whole
runaround you have to do for launchd to be able to access `~/Downloads`;
something about giving `/bin/bash` full disk access, which doesn't seem
worth doing (or particularly safe, for that matter…? 🤷🙄).

Anyway, I'm considering just having a LaunchAgent display a daily
reminder to do it, or triggering it on startup some other way; there are
probably a bunch of options I haven't even come up with yet.