Start setting up for themux directory structure
As mentioned in the recent addition(s) to the README. For now it's juts
`qnd_setup.sh`, which is just a quick and dirty little symlinking
script; I'm not sure if there's a standard tool that can handle that or
if I should just go with what I have already.

Also, added a deps command.