Make some small improvments
Namely, not running the same thing over and over if we can avoid doing
so. There's still some improvement to probably be had by just using a
single ssh-agent for an entire tmux session instead of spawning a new
one for every single tmux pane, but there's some tinkering to get that
to work apparently. In the meantime, one per pane works just fine.